feat: Refactor code to use FerryService for creating Ferries
The code changes introduce a new `FerryService` trait and a corresponding implementation `FerryServiceImpl`. The `FerryService` trait defines a `create` method for creating Ferries with a specified capacity. The `FerryServiceImpl` implementation generates a unique ID for each Ferry and assigns a random name using the `CompanyName` faker. This refactoring allows for better separation of concerns and improves the extensibility of the codebase.

+++ producer/src/lib/domain/models.rs
@@ -2,8 +2,9 @@
 pub struct Ferry {
     pub id: String,
     pub name: String,
-    pub capacity: u8,
+    pub capacity: i32,
 }
+
 
 #[derive(Debug, serde::Serialize)]
 pub struct Car {
@@ -11,7 +12,7 @@ pub struct Car {
     pub licence_plate: String,
     pub brand: String,
     pub color: String,
-    pub capacity: u8,
+    pub capacity: i32,
 }
 
 #[derive(Debug, serde::Serialize)]
